move api calls to separate module
